Añado lógicas para seguridad.

+++ backend/src/main/java/com/example/services/CarritoService.java
@@ -6,7 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.example.model.Carrito;
+import com.example.model.Usuario;
 import com.example.repository.CarritoRepository;
+import com.example.security.ForbiddenException;
+import com.example.security.NotFoundException;
 
 @Service
 public class CarritoService {
@@ -18,11 +21,22 @@ public class CarritoService {
         return repoCarrito.findByUsuario_idUsuario(id);
     }
 
-    public Carrito addCarrito(Carrito carrito) {
-        return repoCarrito.save(carrito);
+    public Carrito addCarrito(Carrito carrito, Usuario usuario) {
+        if (usuario.getIdUsuario() == carrito.getUsuario().getIdUsuario()) {
+            return repoCarrito.save(carrito);
+        } else {
+            throw new ForbiddenException();
+        }
     }
 
-    public void borrarCarrito(Integer id) {
+    public void borrarCarrito(Integer id, Usuario usuario) {
+        Carrito carrito = repoCarrito.findById(id).orElse(null);
+        if (carrito == null) {
+            throw new NotFoundException();
+        }
+        if (carrito.getUsuario().getIdUsuario() != usuario.getIdUsuario()) {
+            throw new ForbiddenException();
+        }
         repoCarrito.deleteById(id);
     }
